requirements.txt
pandas>=2.0
numpy>=1.24
biopython>=1.83
pyyaml>=6.0
lxml>=5.2
